Separate query execution from Semantic Search window UI
