+++ src/VisualStudio/CSharp/Impl/SemanticSearch/SemanticSearchToolWindowImpl.cs
@@ -3,7 +3,6 @@
 // See the LICENSE file in the project root for more information.
 
 using System;
-using System.Collections.Immutable;
 using System.Composition;
 using System.Diagnostics.CodeAnalysis;
 using System.Threading;
@@ -15,20 +14,15 @@ using System.Windows.Data;
 using System.Windows.Markup;
 using System.Windows.Media;
 using Microsoft.CodeAnalysis;
-using Microsoft.CodeAnalysis.Classification;
 using Microsoft.CodeAnalysis.Editor.Host;
 using Microsoft.CodeAnalysis.Editor.Shared.Utilities;
 using Microsoft.CodeAnalysis.ErrorReporting;
-using Microsoft.CodeAnalysis.FindUsages;
 using Microsoft.CodeAnalysis.Host;
 using Microsoft.CodeAnalysis.Host.Mef;
-using Microsoft.CodeAnalysis.Internal.Log;
 using Microsoft.CodeAnalysis.Navigation;
-using Microsoft.CodeAnalysis.Notification;
 using Microsoft.CodeAnalysis.Options;
 using Microsoft.CodeAnalysis.SemanticSearch;
 using Microsoft.CodeAnalysis.Shared.TestHooks;
-using Microsoft.CodeAnalysis.Threading;
 using Microsoft.VisualStudio.Editor;
 using Microsoft.VisualStudio.Extensibility.VSSdkCompatibility;
 using Microsoft.VisualStudio.Imaging;
@@ -54,7 +48,7 @@ using TextSpan = Microsoft.CodeAnalysis.Text.TextSpan;
 [Export(typeof(SemanticSearchToolWindowImpl))]
 [method: ImportingConstructor]
 [method: Obsolete(MefConstruction.ImportingConstructorMessage, error: true)]
-internal sealed class SemanticSearchToolWindowImpl(
+internal sealed partial class SemanticSearchToolWindowImpl(
     IHostWorkspaceProvider hostWorkspaceProvider,
     IThreadingContext threadingContext,
     ITextEditorFactoryService textEditorFactory,
@@ -68,7 +62,7 @@ internal sealed class SemanticSearchToolWindowImpl(
     ITextUndoHistoryRegistry undoHistoryRegistry,
     ISemanticSearchCopilotService copilotService,
     Lazy<ISemanticSearchCopilotUIProvider> copilotUIProvider, // lazy to avoid loading Microsoft.VisualStudio.LanguageServices.ExternalAccess.Copilot
-    IVsService<SVsUIShell, IVsUIShell> vsUIShellProvider) : ISemanticSearchWorkspaceHost, OptionsProvider<ClassificationOptions>, IDisposable
+    IVsService<SVsUIShell, IVsUIShell> vsUIShellProvider) : ISemanticSearchWorkspaceHost, IDisposable
 {
     private const int ToolBarHeight = 26;
     private const int ToolBarButtonSize = 20;
@@ -569,8 +563,6 @@ internal sealed class SemanticSearchToolWindowImpl(
         var querySolution = _semanticSearchWorkspace.Value.CurrentSolution;
         var queryDocument = SemanticSearchUtilities.GetQueryDocument(querySolution);
 
-        var resultsObserver = new ResultsObserver(queryDocument, presenterContext);
-
         var completionToken = _asyncListener.BeginAsyncOperation(nameof(SemanticSearchToolWindow) + ".Execute");
         _ = ExecuteAsync(cancellationSource.Token).ReportNonFatalErrorAsync().CompletesAsyncOperation(completionToken);
 
@@ -578,54 +570,13 @@ internal sealed class SemanticSearchToolWindowImpl(
         {
             await TaskScheduler.Default;
 
-            ExecuteQueryResult result = default;
-
-            var canceled = false;
-            string? queryString = null;
-
             try
             {
-                var solution = workspace.CurrentSolution;
-
-                if (solution.ProjectIds is [])
-                {
-                    await presenterContext.ReportNoResultsAsync(ServicesVSResources.Search_found_no_results_no_csharp_or_vb_projects_opened, cancellationToken).ConfigureAwait(false);
-                }
-                else
-                {
-                    var query = await queryDocument.GetTextAsync(cancellationToken).ConfigureAwait(false);
-                    queryString = query.ToString();
-
-                    result = await RemoteSemanticSearchServiceProxy.ExecuteQueryAsync(
-                        solution,
-                        LanguageNames.CSharp,
-                        queryString,
-                        SemanticSearchUtilities.ReferenceAssembliesDirectory,
-                        resultsObserver,
-                        this,
-                        cancellationToken).ConfigureAwait(false);
-
-                    foreach (var error in result.compilationErrors)
-                    {
-                        await presenterContext.OnDefinitionFoundAsync(new SearchCompilationFailureDefinitionItem(error, queryDocument), cancellationToken).ConfigureAwait(false);
-                    }
-                }
-            }
-            catch (Exception e) when (FatalError.ReportAndPropagateUnlessCanceled(e, cancellationToken, ErrorSeverity.Critical))
-            {
-                result = new ExecuteQueryResult(compilationErrors: [], e.Message);
-            }
-            catch (OperationCanceledException)
-            {
-                result = new ExecuteQueryResult(compilationErrors: [], ServicesVSResources.Search_cancelled);
-                canceled = true;
+                var executor = new SemanticSearchQueryExecutor(presenterContext, globalOptions);
+                await executor.ExecuteAsync(query: null, queryDocument, workspace.CurrentSolution, cancellationToken).ConfigureAwait(false);
             }
             finally
             {
-                // Notify the presenter even if the search has been cancelled.
-                var completionToken = _asyncListener.BeginAsyncOperation(nameof(SemanticSearchToolWindow) + ".Completion");
-                _ = CompleteSearchAsync().ReportNonFatalErrorAsync().CompletesAsyncOperation(completionToken);
-
                 // Only clear pending source if it is the same as our source (otherwise, another execution has already kicked off):
                 Interlocked.CompareExchange(ref _pendingExecutionCancellationSource, value: null, cancellationSource);
 
@@ -638,51 +589,6 @@ internal sealed class SemanticSearchToolWindowImpl(
 
                 // Update UI:
                 UpdateUIState();
-
-                async Task CompleteSearchAsync()
-                {
-                    var errorMessage = result.ErrorMessage;
-
-                    if (errorMessage != null)
-                    {
-                        if (result.ErrorMessageArgs != null)
-                        {
-                            errorMessage = string.Format(errorMessage, result.ErrorMessageArgs);
-                        }
-
-                        await presenterContext.ReportMessageAsync(
-                            errorMessage,
-                            canceled ? NotificationSeverity.Information : NotificationSeverity.Error,
-                            CancellationToken.None).ConfigureAwait(false);
-                    }
-
-                    await presenterContext.OnCompletedAsync(CancellationToken.None).ConfigureAwait(false);
-
-                    if (queryString != null)
-                    {
-                        Logger.Log(FunctionId.SemanticSearch_QueryExecution, KeyValueLogMessage.Create(map =>
-                        {
-                            map["Query"] = new PiiValue(queryString);
-
-                            if (canceled)
-                            {
-                                map["Canceled"] = true;
-                            }
-                            else if (result.ErrorMessage != null)
-                            {
-                                map["ErrorMessage"] = result.ErrorMessage;
-
-                                if (result.ErrorMessageArgs != null)
-                                {
-                                    map["ErrorMessageArgs"] = new PiiValue(string.Join("|", result.ErrorMessageArgs));
-                                }
-                            }
-
-                            map["ExecutionTimeMilliseconds"] = (long)result.ExecutionTime.TotalMilliseconds;
-                            map["EmitTime"] = (long)result.EmitTime.TotalMilliseconds;
-                        }));
-                    }
-                }
             }
         }
     }
@@ -710,30 +616,11 @@ internal sealed class SemanticSearchToolWindowImpl(
             return true;
         });
 
-    public ValueTask<ClassificationOptions> GetOptionsAsync(Microsoft.CodeAnalysis.Host.LanguageServices languageServices, CancellationToken cancellationToken)
-        => new(globalOptions.GetClassificationOptions(languageServices.Language));
-
     private sealed class CopilotUI
     {
         public required FrameworkElement Control { get; init; }
         public required ITextBoxControl Input { get; init; }
         public required ComboBox ModelPicker { get; init; }
-    }
-
-    private sealed class ResultsObserver(Document queryDocument, IFindUsagesContext presenterContext) : ISemanticSearchResultsObserver
-    {
-        public ValueTask OnDefinitionFoundAsync(DefinitionItem definition, CancellationToken cancellationToken)
-            => presenterContext.OnDefinitionFoundAsync(definition, cancellationToken);
-
-        public ValueTask AddItemsAsync(int itemCount, CancellationToken cancellationToken)
-            => presenterContext.ProgressTracker.AddItemsAsync(itemCount, cancellationToken);
-
-        public ValueTask ItemsCompletedAsync(int itemCount, CancellationToken cancellationToken)
-            => presenterContext.ProgressTracker.ItemsCompletedAsync(itemCount, cancellationToken);
-
-        public ValueTask OnUserCodeExceptionAsync(UserCodeExceptionInfo exception, CancellationToken cancellationToken)
-            => presenterContext.OnDefinitionFoundAsync(
-                new SearchExceptionDefinitionItem(exception.Message, exception.TypeName, exception.StackTrace, new DocumentSpan(queryDocument, exception.Span)), cancellationToken);
     }
 
     private sealed class CommandFilter : IOleCommandTarget
