+++ src/VisualStudio/Core/Def/SemanticSearch/ISemanticSearchPresenterController.cs
@@ -1,0 +1,13 @@
+﻿// Licensed to the .NET Foundation under one or more agreements.
+// The .NET Foundation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System.Threading;
+using System.Threading.Tasks;
+
+namespace Microsoft.CodeAnalysis.SemanticSearch;
+
+internal interface ISemanticSearchPresenterController
+{
+    Task ExecuteQueryAsync(string query, CancellationToken cancellationToken);
+}
